Moved functions from 'match_graphs.py' to graph.py
