+++ datamol/__init__.py
@@ -127,5 +127,5 @@ from .mcs import find_mcs
 
 from .graph import to_graph
 from .graph import get_all_path_between
-from .match_graphs import match_molecular_graphs
-from .match_graphs import reorder_mol_from_template
+from .graph import match_molecular_graphs
+from .graph import reorder_mol_from_template
